+++ cmd/operator-sdk/internal/genutil/openapi.go
@@ -15,8 +15,9 @@
 package genutil
 
 import (
+	"flag"
 	"fmt"
-	"os/exec"
+	"os"
 	"path/filepath"
 	"strings"
 
@@ -25,22 +26,19 @@ import (
 	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	gengoargs "k8s.io/gengo/args"
+	generatorargs "k8s.io/kube-openapi/cmd/openapi-gen/args"
+	"k8s.io/kube-openapi/pkg/generators"
 )
 
-// OpenAPIGen generates OpenAPI validation specs for all CRD's in dirs. hf is
-// a path to a header file containing text to add to generated files.
-func OpenAPIGen(hf string) error {
+// OpenAPIGen generates OpenAPI validation specs for all CRD's in dirs.
+func OpenAPIGen() error {
 	projutil.MustInProjectRoot()
 
 	absProjectPath := projutil.MustGetwd()
 	repoPkg := projutil.CheckAndGetProjectGoPkg()
-	srcDir := filepath.Join(absProjectPath, "vendor", "k8s.io", "kube-openapi")
-	binDir := filepath.Join(absProjectPath, scaffold.BuildBinDir)
-
-	if err := buildOpenAPIGenBinary(binDir, srcDir); err != nil {
-		return err
-	}
 
 	gvMap, err := parseGroupVersions()
 	if err != nil {
@@ -54,10 +52,9 @@ func OpenAPIGen(hf string) error {
 	log.Infof("Running OpenAPI code-generation for Custom Resource group versions: [%v]\n", gvb.String())
 
 	apisPkg := filepath.Join(repoPkg, scaffold.ApisDir)
-	fqApiStr := createFQApis(apisPkg, gvMap)
-	fqApis := strings.Split(fqApiStr, ",")
-	f := func(a string) error { return openAPIGen(binDir, a, fqApis) }
-	if err = withHeaderFile(hf, f); err != nil {
+	fqApis := createFQAPIs(apisPkg, gvMap)
+	f := func(a string) error { return openAPIGen(a, fqApis) }
+	if err = generateWithHeaderFile(f); err != nil {
 		return err
 	}
 
@@ -96,25 +93,37 @@ func OpenAPIGen(hf string) error {
 	return nil
 }
 
-func buildOpenAPIGenBinary(binDir, codegenSrcDir string) error {
-	genDirs := []string{"./cmd/openapi-gen"}
-	return buildCodegenBinaries(genDirs, binDir, codegenSrcDir)
-}
-
-func openAPIGen(binDir, hf string, fqApis []string) (err error) {
-	cgPath := filepath.Join(binDir, "openapi-gen")
-	for _, fqApi := range fqApis {
-		args := []string{
-			"--input-dirs", fqApi,
-			"--output-package", fqApi,
-			"--output-file-base", "zz_generated.openapi",
-			// openapi-gen requires a boilerplate file. Either use header or an
-			// empty file if header is empty.
-			"--go-header-file", hf,
+func openAPIGen(hf string, fqApis []string) error {
+	wd, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+	flag.Set("logtostderr", "true")
+	for _, api := range fqApis {
+		api = filepath.FromSlash(api)
+		// Use relative API path so the generator writes to the correct path.
+		apiPath := "." + string(filepath.Separator) + api[strings.Index(api, scaffold.ApisDir):]
+		args := &gengoargs.GeneratorArgs{
+			InputDirs:          []string{apiPath},
+			OutputFileBaseName: "zz_generated.openapi",
+			OutputPackagePath:  filepath.Join(wd, apiPath),
+			GoHeaderFilePath:   hf,
+			CustomArgs: &generatorargs.CustomArgs{
+				// Print API rule violations to stdout
+				ReportFilename: "-",
+			},
 		}
-		cmd := exec.Command(cgPath, args...)
-		if err = projutil.ExecCmd(cmd); err != nil {
-			return fmt.Errorf("failed to perform openapi code-generation: %v", err)
+		if err := generatorargs.Validate(args); err != nil {
+			return errors.Wrap(err, "openapi-gen argument validation error")
+		}
+
+		err := args.Execute(
+			generators.NameSystems(),
+			generators.DefaultNameSystem(),
+			generators.Packages,
+		)
+		if err != nil {
+			return errors.Wrap(err, "openapi-gen generator error")
 		}
 	}
 	return nil