+++ cmd/operator-sdk/test/local.go
@@ -40,16 +40,17 @@ import (
 var deployTestDir = filepath.Join(scaffold.DeployDir, "test")
 
 type testLocalConfig struct {
-	kubeconfig        string
-	globalManPath     string
-	namespacedManPath string
-	goTestFlags       string
-	moleculeTestFlags string
-	namespace         string
-	upLocal           bool
-	noSetup           bool
-	debug             bool
-	image             string
+	kubeconfig         string
+	globalManPath      string
+	namespacedManPath  string
+	goTestFlags        string
+	moleculeTestFlags  string
+	namespace          string
+	upLocal            bool
+	noSetup            bool
+	debug              bool
+	image              string
+	localOperatorFlags string
 }
 
 var tlConfig testLocalConfig
@@ -70,6 +71,7 @@ func newTestLocalCmd() *cobra.Command {
 	testCmd.Flags().BoolVar(&tlConfig.noSetup, "no-setup", false, "Disable test resource creation")
 	testCmd.Flags().BoolVar(&tlConfig.debug, "debug", false, "Enable debug-level logging")
 	testCmd.Flags().StringVar(&tlConfig.image, "image", "", "Use a different operator image from the one specified in the namespaced manifest")
+	testCmd.Flags().StringVar(&tlConfig.localOperatorFlags, "local-operator-flags", "", "The flags that the operator needs (while using --up-local). Example: \"--flag1 value1 --flag2=value2\"")
 
 	return testCmd
 }
@@ -207,12 +209,16 @@ func testLocalGoFunc(cmd *cobra.Command, args []string) error {
 	}
 	if tlConfig.upLocal {
 		testArgs = append(testArgs, "-"+test.LocalOperatorFlag)
+		if tlConfig.localOperatorFlags != "" {
+			testArgs = append(testArgs, "-"+test.LocalOperatorArgs, tlConfig.localOperatorFlags)
+		}
 	}
 	opts := projutil.GoTestOptions{
 		GoCmdOptions: projutil.GoCmdOptions{
 			PackagePath: args[0] + "/...",
 			Env:         append(os.Environ(), fmt.Sprintf("%v=%v", test.TestNamespaceEnv, tlConfig.namespace)),
 			Dir:         projutil.MustGetwd(),
+			GoMod:       projutil.IsDepManagerGoMod(),
 		},
 		TestBinaryArgs: testArgs,
 	}