+++ internal/pkg/scaffold/ansible/deploy_operator.go
@@ -25,7 +25,6 @@ const DeployOperatorFile = "operator.yaml"
 
 type DeployOperator struct {
 	input.Input
-	IsClusterScoped bool
 }
 
 // GetInput - gets the input
@@ -76,13 +75,9 @@ spec:
             name: runner
           env:
             - name: WATCH_NAMESPACE
-              [[- if .IsClusterScoped ]]
-              value: ""
-              [[- else ]]
               valueFrom:
                 fieldRef:
                   fieldPath: metadata.namespace
-              [[- end]]
             - name: POD_NAME
               valueFrom:
                 fieldRef: